requirements.txt
python == 3.9.15
pytorch == 1.12.1
torchvision == 0.13.1
einops == 0.6.0
tensorboardX == 2.5.1
numpy == 1.23.4
pickle
yaml
argparse